refactor topics module path
