--- feature-topics/src/main/java/st/slex/feature_topics/domain/TopicsInteractor.kt
@@ -1,9 +1,0 @@
-package st.slex.feature_topics.domain
-
-import androidx.paging.PagingData
-import kotlinx.coroutines.flow.Flow
-import st.slex.feature_topics.domain.model.TopicsUIModel
-
-interface TopicsInteractor {
-    val topics: Flow<PagingData<TopicsUIModel>>
-}
