--- feature-topics/src/main/java/st/slex/feature_topics/domain/model/Mapper.kt
@@ -1,10 +1,0 @@
-package st.slex.feature_topics.domain.model
-
-import st.slex.feature_topics.data.model.TopicsModel
-
-fun TopicsModel.toTopicsUI() = TopicsUIModel(
-    id = id,
-    title = title,
-    url = coverPhoto.urls.regular,
-    description = description
-)
