--- feature-topics/src/main/java/st/slex/feature_topics/domain/model/TopicsUIModel.kt
@@ -1,8 +1,0 @@
-package st.slex.feature_topics.domain.model
-
-data class TopicsUIModel(
-    val id: String,
-    val title: String,
-    val url: String,
-    val description: String
-)
