--- feature-topics/src/main/java/st/slex/feature_topics/ui/components/TopicBox.kt
@@ -1,35 +1,0 @@
-package st.slex.feature_topics.ui.components
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.padding
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import st.slex.feature_topics.domain.model.TopicsUIModel
-
-@Composable
-fun TopicBox(
-    item: TopicsUIModel?,
-    onBoxClicked: () -> Unit,
-    isClicked: Boolean,
-    modifier: Modifier = Modifier,
-) {
-    Box(
-        modifier = modifier
-            .padding(top = 32.dp)
-            .clickable(onClick = onBoxClicked),
-    ) {
-        TopicsImage(
-            isSelectedState = isClicked,
-            url = item?.url.orEmpty(),
-            id = item?.id.orEmpty()
-        )
-        TopicsDescription(
-            modifier = Modifier.align(Alignment.BottomStart),
-            isSelectedState = isClicked,
-            description = item?.description.orEmpty()
-        )
-    }
-}
