+++ feature/topics/src/main/AndroidManifest.xml
