+++ feature/topics/src/main/java/com/stslex/csplashscreen/feature/topics/domain/TopicsInteractor.kt
@@ -1,0 +1,9 @@
+package com.stslex.csplashscreen.feature.topics.domain
+
+import androidx.paging.PagingData
+import kotlinx.coroutines.flow.Flow
+import com.stslex.csplashscreen.feature.topics.domain.model.TopicsUIModel
+
+interface TopicsInteractor {
+    val topics: Flow<PagingData<TopicsUIModel>>
+}
