+++ feature/topics/src/main/java/com/stslex/csplashscreen/feature/topics/domain/model/Mapper.kt
@@ -1,0 +1,10 @@
+package com.stslex.csplashscreen.feature.topics.domain.model
+
+import com.stslex.csplashscreen.feature.topics.data.model.TopicsModel
+
+fun TopicsModel.toTopicsUI() = TopicsUIModel(
+    id = id,
+    title = title,
+    url = coverPhoto.urls.regular,
+    description = description
+)
