+++ feature/topics/src/main/java/com/stslex/csplashscreen/feature/topics/domain/model/TopicsUIModel.kt
@@ -1,0 +1,8 @@
+package com.stslex.csplashscreen.feature.topics.domain.model
+
+data class TopicsUIModel(
+    val id: String,
+    val title: String,
+    val url: String,
+    val description: String
+)
