+++ settings.gradle.kts
@@ -32,5 +32,5 @@ include(":feature:collection")
 include(":feature:user")
 include(":feature-photo-detail")
 include(":feature-image-raw")
-include(":feature-topics")
+include(":feature:topics")
 include(":feature:search")
